requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
itsdangerous==2.1.2
posthog==3.0.0
stripe==7.14.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
Flask-Bootstrap==3.3.7.1

# Note: If installation fails, try:
# 1. Install Microsoft Visual C++ Build Tools
# 2. Or use: pip install psycopg2-binary --only-binary :all: 